requirements.txt
Django==5.2.3
gunicorn
python-dotenv
weasyprint
requests
python-dateutil
dj-database-url
psycopg2-binary
django-payfast
whitenoise
boto3
django-storages